Edit Meeting Commit (#120)
* Refactor MeetingForm. Create NewMeeting

* Implement Edit Meeting

* Parsing existing meetings information to TimePicker and Dropdown

* Clean test page
