+++ frontend/app/components/templates/HomePageLayout.tsx
@@ -4,73 +4,23 @@ import styles from "../../../styles/HomePageLayout.module.scss";
 import CalendarNavbar from "../organisms/CalendarNavbar";
 import CalendarSidebar from "../organisms/CalendarSidebar";
 import ViewMeetingDetails from "../organisms/ViewMeeting";
+import EditMeetingSidebar from "../organisms/EditMeeting";
 import DailyView from "../organisms/DailyView";
 import { convertUTCToET } from "../../../util/timeUtils";
-
-type MeetingDetails = {
-  id: string;
-  mid: string;
-  title: string;
-  description?: string;
-  creator: string;
-  group: string;
-  startDateTime: Date;
-  endDateTime: Date;
-  zoomAccount?: string;
-  zoomLink?: string;
-  zid?: string;
-  type: string;
-  room: string;
-  recurrence?: string;
-};
+import { IMeeting } from "../../../util/models";
 
 const HomePage = () => {
   const [selectedDate, setSelectedDate] = useState<Date>(new Date());
-  const [selectedMeeting, setSelectedMeeting] = useState<MeetingDetails | null>(null);
+  const [selectedMeeting, setSelectedMeeting] = useState<IMeeting | null>(null);
   const [selectedMeetingID, setSelectedMeetingID] = useState<string | null>(null);
   const [selectedNewMeeting, setSelectedNewMeeting] = useState<boolean | null>(false);
-  const [inputMeetingTitleValue, setMeetingTitleValue] = useState(""); // Meeting title
-  const [dateValue, setDateValue] = useState<string>(""); // Initial date value as empty
-  const [timeValue, setTimeValue] = useState<string>(""); // Initial time range as empty
-  const [freqValue, setFreqValue] = useState<string>("Never"); // Default frequency value
-  const [inputEmailValue, setEmailValue] = useState(""); // Email input value
-  const [inputDescriptionValue, setDescriptionValue] = useState(""); // Description input value
-
-  const roomOptions = [
-    "Serenity Room",
-    "Seeds of Hope",
-    "Unity Room",
-    "Room for Improvement",
-    "Small but Powerful - Right",
-    "Small but Powerful - Left",
-  ];
-
-  const meetingTypeOptions = [
-    "AA",
-    "Al-Anon",
-    "Other",
-  ];
-
-  const zoomAccountOptions = [
-    "Zoom Email 1",
-    "Zoom Email 2",
-    "Zoom Email 3",
-    "Zoom Email 4",
-  ];
-
-  const handleFileSelect = (file: File | null) => {
-    if (file) {
-      console.log("File selected:", file);
-    } else {
-      console.log("No file selected");
-    }
-  };
+  const [showEditMeeting, setShowEditMeeting] = useState(false);
 
   const fetchMeetingDetails = async (meetingId: string) => {
     try {
       const response = await fetch(`/api/retrieve/meeting/${meetingId}`, { method: 'GET' });
       if (response.ok) {
-        const data: MeetingDetails = await response.json(); // Ensure data matches MeetingDetails type
+        const data: IMeeting = await response.json(); // Ensure data matches MeetingDetails type
         setSelectedMeeting(data);
       } else {
         console.error("Failed to fetch meeting details");
@@ -94,29 +44,43 @@ const HomePage = () => {
     setSelectedNewMeeting(false);
   };
 
-  const handleEdit = () => console.log("Edit meeting");
+  const handleOpenEdit = () => {
+    setShowEditMeeting(true);
+  };
+
+  const handleCloseEdit = () => {
+    setShowEditMeeting(false);
+  };
+
+  const handleCloseNewMeeting = () => {
+    setSelectedNewMeeting(false);
+  };
 
   const handleDelete = async (mid: string) => {
     try {
-      const response = await fetch("/api/delete/meeting", {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ mid }),
+      const response = await fetch('/api/delete/meeting', {
+        method: 'DELETE',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          mid
+        }),
       });
 
       if (!response.ok) {
-        alert("Error: Unsuccessful delete");
+        alert("Error : Unsuccessful delete")
         throw new Error(`HTTP error! status: ${response.status}`);
       }
-
       setSelectedMeeting(null);
-      alert("Meeting deleted successfully!");
+
+      const meetingResponse = await response.json();
+      console.log(meetingResponse);
+      alert("Meeting deleted successfully! Please check the Meeting collection on MongoDB.")
+
     } catch (error) {
-      console.error("Error deleting the meeting:", error);
+      console.error('There was an error fetching the data:', error);
     }
-  };
-  const handleCloseNewMeeting = () => {
-    setSelectedNewMeeting(false);
   };
 
   const [filters, setFilters] = useState({
@@ -161,49 +125,63 @@ const HomePage = () => {
   return (
     <div className={styles.container}>
       <div className={styles.sidebar}>
-        {selectedMeeting ? (
-          <ViewMeetingDetails
-            id={selectedMeeting.id}
-            mid={selectedMeeting.mid}
-            title={selectedMeeting.title}
-            description={selectedMeeting.description}
-            creator={selectedMeeting.creator}
-            group={selectedMeeting.group}
+        {showEditMeeting && selectedMeeting ? (
+            // TODO: Update Edit Meeting component
+            <EditMeetingSidebar
+              meeting={selectedMeeting}
+              onClose={handleCloseEdit}
+              onUpdateSuccess={() => {
+                console.log("Meeting updated!");
+                // Refresh the meeting data after successful update
+                if (selectedMeeting.mid) {
+                  fetchMeetingDetails(selectedMeeting.mid);
+                }
+              }}
+              />) :
+          selectedMeeting ? (
+            // TODO: Check Handle Delete in ViewMeetingDetails
+            <ViewMeetingDetails
+              mid={selectedMeeting.mid}
+              title={selectedMeeting.title}
+              description={selectedMeeting.description}
+              creator={selectedMeeting.creator}
+              group={selectedMeeting.group}
 
-            startDateTime={convertESTStringToDate(
-              convertUTCToET(
-                selectedMeeting.startDateTime instanceof Date
-                  ? selectedMeeting.startDateTime.toISOString()
-                  : selectedMeeting.startDateTime
-              )
-            )}
-            
-            endDateTime={convertESTStringToDate(
-              convertUTCToET(
-                selectedMeeting.endDateTime instanceof Date
-                  ? selectedMeeting.endDateTime.toISOString()
-                  : selectedMeeting.endDateTime
-              )
-            )}
+              startDateTime={convertESTStringToDate(
+                convertUTCToET(
+                  selectedMeeting.startDateTime instanceof Date
+                    ? selectedMeeting.startDateTime.toISOString()
+                    : selectedMeeting.startDateTime
+                )
+              )}
+              
+              endDateTime={convertESTStringToDate(
+                convertUTCToET(
+                  selectedMeeting.endDateTime instanceof Date
+                    ? selectedMeeting.endDateTime.toISOString()
+                    : selectedMeeting.endDateTime
+                )
+              )}
 
-            zoomAccount={selectedMeeting.zoomAccount}
-            zoomLink={selectedMeeting.zoomLink}
-            zid={selectedMeeting.zid}
-            type={selectedMeeting.type}
-            room={selectedMeeting.room}
-            recurrence={selectedMeeting.recurrence}
-            onBack={handleBack}
-            onEdit={handleEdit}
-            onDelete={() => handleDelete(selectedMeeting.mid)}
-          />
-        ) : (
-          <CalendarSidebar 
-            filters={filters}
-            setFilters={setFilters}
-            selectedDate={selectedDate} 
-            setSelectedDate={setSelectedDate} />
-        )}
+              zoomAccount={selectedMeeting.zoomAccount}
+              zoomLink={selectedMeeting.zoomLink}
+              zid={selectedMeeting.zid}
+              type={selectedMeeting.type}
+              room={selectedMeeting.room}
+              // recurrence={selectedMeeting.recurrence} // TODO: Update when merge with Recurring Meetings
+              onBack={handleBack}
+              onEdit={handleOpenEdit}
+              onDelete={handleDelete}
+            />
+          ) : (
+            <CalendarSidebar 
+              filters={filters}
+              setFilters={setFilters}
+              selectedDate={selectedDate} 
+              setSelectedDate={setSelectedDate} />
+          )}
       </div>
+
       <div className={styles.primaryCalendar}>
         <CalendarNavbar
           selectedDate={selectedDate}
