+++ frontend/styles/components/organisms/CalendarSidebar.module.scss
@@ -7,20 +7,3 @@
   background: #FFF;
   color: rgb(0, 0, 0);
 }
-
-.newMeetingHeader {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  color: var(--Dark-Navy, #3A3E49);
-  font-family: "Source Sans Pro";
-  font-style: normal;
-  font-weight: 600;
-  background-color: white;
-}
-
-.iconButton {
-  width: auto;
-  height: 40px;
-}
