+++ frontend/styles/components/organisms/MeetingForm.module.scss
@@ -60,3 +60,20 @@
     padding: 20px;
   }
 }
+
+.meetingHeader {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  color: var(--Dark-Navy, #3A3E49);
+  font-family: "Source Sans Pro";
+  font-style: normal;
+  font-weight: 600;
+  background-color: white;
+}
+
+.iconButton {
+  width: auto;
+  height: 40px;
+}
